Cast sales invoice RPC customer id to integer

+++ supabase/sql/2026-05-27_save_sales_invoice_rpc.sql
@@ -46,7 +46,7 @@ begin
   )
   values (
     v_id,
-    nullif(p_invoice->>'customer_id', ''),
+    nullif(p_invoice->>'customer_id', '')::integer,
     coalesce(nullif(p_invoice->>'customer_name', ''), 'Unknown'),
     nullif(p_invoice->>'reference', ''),
     nullif(p_invoice->>'vehicle_number', ''),
